backpedaling a bit on ScalaTest version since 2.0M6-SNAP35 is newest
available for 2.11.0-M4

+++ nightly.sh
@@ -3,7 +3,7 @@
 # -e makes the whole thing die with an error if any command does
 # -v lets you see the commands as they happen
 
-if [ "$1" != --noclean ]; then
+if [ "$1" != --noclean ] && [ "$1" != --no-clean ] ; then
   git clean -fdX
   git submodule update --init
   git submodule foreach git clean -fdX
@@ -32,11 +32,18 @@ echo "*** done: nogen fast:test"
 if [ ${PIPESTATUS[0]} -ne 0 ] ; then echo "*** FAILED: sbt all"; exit 1; fi
 echo "*** done: sbt all"
 
-./sbt slow:test 2>&1 | tee tmp/nightly/3-slow-test.txt
+# experimenting hoping that intermittent Travis failures will stop.
+# hopefully turning off logBuffered for the slow tests means that we
+# won't go 10 minutes without producing any output.  with logBuffered
+# off, the output from the different tests gets scrambled together,
+# but we can't turn off parallelExecution or the whole thing will
+# take more than 50 minutes to run (the Travis limit) - ST 8/29/13
+
+./sbt 'set logBuffered in test := false' slow:test 2>&1 | tee tmp/nightly/3-slow-test.txt
 if [ ${PIPESTATUS[0]} -ne 0 ] ; then echo "*** FAILED: slow:test"; exit 1; fi
 echo "*** done: slow:test"
 
-./sbt nogen slow:test 2>&1 | tee tmp/nightly/4-nogen-slow-test.txt
+./sbt 'set logBuffered in test := false' nogen slow:test 2>&1 | tee tmp/nightly/4-nogen-slow-test.txt
 if [ ${PIPESTATUS[0]} -ne 0 ] ; then echo "*** FAILED: nogen slow:test"; exit 1; fi
 echo "*** done: nogen slow:test"
 
@@ -44,9 +51,8 @@ echo "*** done: nogen slow:test"
 if [ ${PIPESTATUS[0]} -ne 0 ] ; then echo "*** FAILED: depend"; exit 1; fi
 echo "*** done: depend"
 
-# commenting out until plugin is available for sbt 0.13 - ST 8/11/13
-#./sbt scalastyle 2>&1 | tee tmp/nightly/6-scalastyle.txt
-#if [ `wc -l < target/scalastyle-result.xml` -ne 2 ] ; then echo "*** FAILED: scalastyle"; exit 1; fi
-#echo "*** done: scalastyle"
+./sbt scalastyle 2>&1 | tee tmp/nightly/6-scalastyle.txt
+if [ `wc -l < target/scalastyle-result.xml` -ne 2 ] ; then echo "*** FAILED: scalastyle"; exit 1; fi
+echo "*** done: scalastyle"
 
 echo "****** all done!"
